Convert repository methods to async and clean up codebase
Refactored multiple repository methods to use async to improve non-blocking performance. Removed Tokio runtime creation where unnecessary and integrated the `async-trait` crate. Additionally, standardized formatting, adjusted `volume` type to `u64`, added cleanup to reduce redundancy, and streamlined function implementations and tests for consistency.

+++ src/infrastructure/clickhouse/model.rs
@@ -19,7 +19,7 @@ pub struct OHLCVData {
     /// Closing price
     pub close: Positive,
     /// Volume traded
-    pub volume: u32,
+    pub volume: u64,
 }
 
 impl fmt::Display for OHLCVData {
